sailtrack/requirements.txt
gpiozero~=1.6.2
paho-mqtt~=1.6.1
smbus2~=0.4.1
timeloop~=1.0.2
